the return of positive random numbers

+++ src/math/GaloisLFSR.hpp
@@ -60,7 +60,7 @@ public:
 	}
 
 	double UNorm() noexcept {
-		return double(Next<std::uint64_t>()) * (1.0 / double(std::numeric_limits<std::uint64_t>::max()));
+		return double(Next<std::uint32_t>()) * (1.0 / double(std::numeric_limits<std::uint32_t>::max()));
 	}
 
 	unsigned int UInt(unsigned int below) noexcept {
